refactor: 회원가입 예외 도메인 하위 클래스로 이동
- 공통 예외가 아니기 때문에 회원 도메인 하위로 이동
