+++ src/test/java/grepp/NBE5_6_2_Team03/api/controller/user/service/UserServiceTest.java
@@ -4,7 +4,7 @@ import grepp.NBE5_6_2_Team03.api.controller.user.dto.request.UserSignUpRequest;
 import grepp.NBE5_6_2_Team03.domain.user.User;
 import grepp.NBE5_6_2_Team03.domain.user.repository.UserRepository;
 import grepp.NBE5_6_2_Team03.domain.user.service.UserService;
-import grepp.NBE5_6_2_Team03.global.exception.UserSignUpException;
+import grepp.NBE5_6_2_Team03.domain.user.exception.UserSignUpException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
